Collected prefetching code in one place and broke dependency from router on it
