+++ runtime/src/app/router/find_anchor.ts
@@ -1,0 +1,4 @@
+export default function find_anchor(node: Node) {
+	while (node && node.nodeName.toUpperCase() !== 'A') node = node.parentNode; // SVG <a> elements have a lowercase name
+	return node;
+}
